+++ packages/design-system/src/components/OcTextInput/OcTextInput.spec.ts
@@ -1,34 +1,10 @@
 import { shallowMount, mount, defaultPlugins } from 'web-test-helpers'
 import OcTextInput from './OcTextInput.vue'
+import { PasswordPolicy } from './../../helpers'
+import { mock } from 'jest-mock-extended'
 
 const defaultProps = {
   label: 'label'
-}
-
-class basicPasswordPolicy {
-  verified = false
-
-  rules = [{ minLength: {} }]
-  check(password) {
-    const result = password.length >= 8
-    this.verified = result
-    return result
-  }
-
-  missing(password) {
-    const result = this.check(password)
-    this.verified = result
-    return {
-      rules: [
-        {
-          code: 'minLength',
-          message: 'At least %{param1} characters',
-          format: ['8'],
-          verified: result
-        }
-      ]
-    }
-  }
 }
 
 Object.assign(navigator, {
@@ -144,15 +120,39 @@ describe('OcTextInput', () => {
         expect(wrapper.emitted('passwordChallengeCompleted')).toBeTruthy()
       })
       it('displays error state if password does not match criteria', async () => {
+        const passwordPolicyMock = mock<PasswordPolicy>()
+        passwordPolicyMock.missing.mockReturnValue({
+          rules: [
+            {
+              code: 'minLength',
+              message: 'At least %{param1} characters',
+              format: ['8'],
+              verified: false
+            }
+          ]
+        })
+        passwordPolicyMock.check.mockReturnValue(false)
         const wrapper = getMountedWrapper({
-          props: { type: 'password', passwordPolicy: new basicPasswordPolicy() }
+          props: { type: 'password', passwordPolicy: passwordPolicyMock }
         })
         await wrapper.find(selectors.inputField).setValue('pass')
         expect(wrapper.html()).toMatchSnapshot()
       })
       it('displays success state if password matches criteria', async () => {
+        const passwordPolicyMock = mock<PasswordPolicy>()
+        passwordPolicyMock.missing.mockReturnValue({
+          rules: [
+            {
+              code: 'minLength',
+              message: 'At least %{param1} characters',
+              format: ['8'],
+              verified: true
+            }
+          ]
+        })
+        passwordPolicyMock.check.mockReturnValue(true)
         const wrapper = getMountedWrapper({
-          props: { type: 'password', passwordPolicy: new basicPasswordPolicy() }
+          props: { type: 'password', passwordPolicy: passwordPolicyMock }
         })
         await wrapper.find(selectors.inputField).setValue('password123')
         expect(wrapper.html()).toMatchSnapshot()
