+++ packages/design-system/src/helpers/types.ts
@@ -26,3 +26,18 @@ export interface ContextualHelper {
   isEnabled: boolean
   data: ContextualHelperData
 }
+
+export interface PasswordPolicy {
+  rules: unknown[]
+
+  check(password: string): boolean
+
+  missing(password: string): {
+    rules: {
+      code: string
+      message: string
+      format: (number | string)[]
+      verified: boolean
+    }[]
+  }
+}
